improvement: split types general
